requirements.txt
streamlit
spacy
nltk
pandas
PyPDF2
python-docx
reportlab
